use identifier daxstudio.daxstudio in winget
The current commands in the documentation don't work for me, so I needed to use an ID. I guess it will be useful for others.

p.s. I didn't use Dax Studio for years, as I wasn't using Windows; however, touching this piece of software again is so pleasant! Dax Studio is solid and awesome! Thank you!